Isolate test to decode request
